show data upon to search or paginate and end search of backEnd
